Adding shiny reset app
I use it to easily keep track of the total attempts I make during shiny hunting a particular pokemon. The app produces files in a manner similar to StreamLabels, which can then be opened up in OBS. I have plans to add more general functionality (such as modifying what the files say and so on, as well as adding a welcome screen and more).
If I get around to it, I'll add the binaries to the release page. Sadly, can't do this from the desktop app. That really should be changed (to allow adding releases from desktop)
I also dabbled in internationalization in this project. Unfortunately, I couldn't get it to work. The "en-GB" is just a dummy locale (technically, it's "english - great britain") that I was using for testing purposes. The satellite assemblies were being built, but, even after changing locales, the data in the en-GB locale wasn't being used for some reason.